requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
binaryornot==0.4.4
chardet==5.2.0
click==8.1.7
cookiecutter==2.6.0
Django==4.2.11
django-admin==2.0.2
django-allauth==0.61.1
django-browser-reload==1.12.1
django-excel-response2==3.0.6
django-six==1.0.5
excel-base==1.0.4
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
isoweek==1.3.3
maxminddb==2.5.2
multidict==6.0.5
oauthlib==3.2.2
pillow==10.2.0
PyJWT==2.8.0
python-slugify==8.0.4
python3-openid==3.2.0
requests-oauthlib==1.3.1
screen==1.0.1
sqlparse==0.4.4
text-unidecode==1.3
TimeConvert==3.0.13
tzlocal==5.2
xlwt==1.3.0
yarl==1.9.4
